客栈传说 v0.0.15 更新说明
1.现在时间跳跃功能移到了最上方的时钟里
2.下调了天2-天5的阳级菜谱数量要求
3.下调了晋升斗速的难度
4.现在在斗技场和公会进行游戏时，也会增加研究菜谱经验
5.现在选择金钱时可以手动输入数值了
6.稍微加快了一些角色的移动速度
7.背包界面增加了一个整理按键，默认按物品类型进行整理
8.现在员工会分散生成在门的附近

+++ Assets/Scrpit/Bean/MVC/Gamble/GambleBaseBean.cs
@@ -9,14 +9,14 @@ public class GambleBaseBean
     public GambleStatusType gambleStatus;
 
     //下注金额
-    public long betForMoneyL;
-    public long betForMoneyM;
-    public long betForMoneyS;
+    public int betForMoneyL;
+    public int betForMoneyM;
+    public int betForMoneyS;
 
     //下注上限
-    public long betMaxForMoneyL;
-    public long betMaxForMoneyM;
-    public long betMaxForMoneyS;
+    public int betMaxForMoneyL;
+    public int betMaxForMoneyM;
+    public int betMaxForMoneyS;
 
     public bool isWin = false;
 
@@ -76,7 +76,7 @@ public class GambleBaseBean
     /// <param name="betMaxForMoneyL"></param>
     /// <param name="betMaxForMoneyM"></param>
     /// <param name="betMaxForMoneyS"></param>
-    public void SetBetMax(long betMaxForMoneyL, long betMaxForMoneyM, long betMaxForMoneyS)
+    public void SetBetMax(int betMaxForMoneyL, int betMaxForMoneyM, int betMaxForMoneyS)
     {
         this.betMaxForMoneyL = betMaxForMoneyL;
         this.betMaxForMoneyM = betMaxForMoneyM;
